pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "contractops"
version = "0.3.0"
description = "CI-grade behavior contracts, baseline replay, and release gates for AI agents."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Ideation Lab" }]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "numpy>=1.26",
]

[project.optional-dependencies]
s3 = ["boto3>=1.34"]
gcs = ["google-cloud-storage>=2.14"]
langchain = ["langchain-core>=0.3"]
anthropic = ["anthropic>=0.40"]
all = [
    "boto3>=1.34",
    "google-cloud-storage>=2.14",
    "langchain-core>=0.3",
    "anthropic>=0.40",
    "httpx>=0.27",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
]
bench = [
    "tabulate>=0.9",
    "rich>=13.0",
]

[project.scripts]
contractops = "contractops.cli:main"

[project.urls]
Repository = "https://github.com/rhajamor/contractops"

[tool.setuptools.packages.find]
include = ["contractops*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
